Split Hermes bridge Python modules
