+++ tests/server/agent-bridge-mcp-tools-filter.test.ts
@@ -28,7 +28,7 @@ import sys
 import threading
 from pathlib import Path
 
-path = Path("packages/server/src/services/hermes/agent-bridge/hermes_bridge.py")
+path = Path("packages/server/src/services/hermes/agent-bridge/python/hermes_bridge.py")
 spec = importlib.util.spec_from_file_location("hermes_bridge", path)
 bridge = importlib.util.module_from_spec(spec)
 sys.modules[spec.name] = bridge
